# docs/swagger.yaml
basePath: /
definitions:
...
  models.FailureResponse:
    description: Standard error response format with status and error message.
    properties:
      error:
        example: Bad Request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  models.RegPatientReq:
    properties:
      address:
        description: |-
          Address is the patient's address.
          required: true
          min length: 5
          max length: 255
        maxLength: 255
        minLength: 5
        type: string
      contactNo:
        description: |-
          ContactNumber is the patient's contact number.
          required: true
          numeric: true
          length: 10 digits
        type: string
      dob:
        description: |-
          DOB is the date of birth of the patient.
          required: true
        type: string
      emergencyName:
        description: |-
          EmergencyName is the name of the emergency contact person.
          required: true
        type: string
      emergencyPhone:
        description: |-
          EmergencyPhone is the phone number of the emergency contact person.
          required: true
          numeric: true
          length: 10 digits
        type: string
      emergencyRelation:
        description: |-
          EmergencyRelation is the relationship to the emergency contact person.
          required: true
        type: string
      fullname:
        description: |-
          FullName is the full name of the patient.
          required: true
          min length: 2
          max length: 100
        maxLength: 100
        minLength: 2
        type: string
      gender:
        description: |-
          Gender represents the patient's gender.
          required: true
          allowed values: MALE, FEMALE, OTHER
        enum:
        - MALE
        - FEMALE
        - OTHER
        type: string
    required:
    - address
    - contactNo
    - dob
    - emergencyName
    - emergencyPhone
    - emergencyRelation
    - fullname
    - gender
    type: object
  models.SigninReq:
    properties:
      email:
        description: |-
          Email is the email address of the user.
          required: true
          format: email
          example: "user@example.com"
        type: string
      password:
        description: |-
          Password is the password chosen by the user.
          required: true
          min length: 8
          max length: 64
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.SignupReq:
    properties:
      activated:
        description: |-
          Activated is a flag that determines whether the user is active.
          optional: true
        type: boolean
      email:
        description: |-
          Email is the email address of the user.
          required: true
          format: email
          example: "user@example.com"
        type: string
      fullname:
        description: |-
          Fullname is the full name of the user.
          required: true
          min length: 3
          max length: 30
        maxLength: 30
        minLength: 3
        type: string
      password:
        description: |-
          Password is the password chosen by the user.
          required: true
          min length: 8
          max length: 64
        maxLength: 64
        minLength: 8
        type: string
      role:
        description: |-
          Role defines the user's role in the system.
          required: true
          allowed values: doctor, receptionist
        enum:
        - doctor
        - receptionist
        type: string
    required:
    - email
    - fullname
    - password
    - role
    type: object
  models.SuccessResponse:
    description: Standard success response format with an optional data field.
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  models.UpdatePatientReq:
    properties:
      address:
        description: |-
          Address is the updated address of the patient.
          optional: true
          min length: 5
          max length: 255
        maxLength: 255
        minLength: 5
        type: string
      age:
        description: |-
          Age is the updated age of the patient.
          optional: true
          numeric: true
          max value: 100
        maximum: 100
        type: integer
      contactNo:
        description: |-
          ContactNumber is the updated contact number of the patient.
          optional: true
          numeric: true
          length: 10 digits
        type: string
      dob:
        description: |-
          DOB is the updated date of birth of the patient.
          optional: true
        type: string
      emergencyName:
        description: |-
          EmergencyName is the updated emergency contact name.
          optional: true
        type: string
      emergencyPhone:
        description: |-
          EmergencyPhone is the updated emergency contact phone number.
          optional: true
          numeric: true
          length: 10 digits
        type: string
      emergencyRelation:
        description: |-
          EmergencyRelation is the updated relationship to the emergency contact.
          optional: true
        type: string
      fullname:
        description: |-
          FullName is the updated full name of the patient.
          optional: true
          min length: 2
          max length: 100
        maxLength: 100
        minLength: 2
        type: string
      gender:
        description: |-
          Gender is the updated gender of the patient.
          optional: true
          allowed values: MALE, FEMALE, OTHER
        enum:
        - MALE
        - FEMALE
        - OTHER
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: codervaidik@gmail.com
    name: Vaidik Bajpai
  description: Backend API for MediBridge, a medical record management system.
  title: MediBridge API
  version: "1.0"
paths:
...
  /v1/patient:
    get:
      consumes:
      - application/json
      description: Lists all registered patients with optional pagination and search.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search term (e.g., name or email)
        in: query
        name: searchTerm
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: List patients
      tags:
      - Patients
    post:
      consumes:
      - application/json
      description: Registers a new patient with the provided details.
      parameters:
      - description: Patient registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegPatientReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Register a new patient
      tags:
      - Patients
  /v1/patient/{patientID}:
    delete:
      consumes:
      - application/json
      description: Deletes a patient by their patient ID.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Delete a patient
      tags:
      - Patients
    get:
      consumes:
      - application/json
      description: Retrieves a patient's details using their patient ID.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Get patient details
      tags:
      - Patients
    put:
      consumes:
      - application/json
      description: Updates details of an existing patient by patient ID.
      parameters:
      - description: Patient ID (UUID)
        in: path
        name: patientID
        required: true
        type: string
      - description: Updated patient data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePatientReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Update a patient
      tags:
      - Patients
...
  /v1/user/signin:
    post:
      consumes:
      - application/json
      description: Authenticates a user and sets a session cookie upon successful
        login.
      parameters:
      - description: Login request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SigninReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Log in a user
      tags:
      - Users
  /v1/user/signup:
    post:
      consumes:
      - application/json
      description: Registers a new user with fullname, email, password, and role.
      parameters:
      - description: Signup request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignupReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureResponse'
      summary: Register a new user
      tags:
      - Users
schemes:
- http
swagger: "2.0"